Day 44
This program defines the read_integer function as previously described. The main function calls read_integer to get a valid integer input from the user and then prints out the entered integer. If the input is invalid, the program will keep prompting the user until a valid integer is provided.